+++ apps/basecamp-fe/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0](https://github.com/yeti-robotics/polar-edge/compare/basecamp-fe-v1.0.1...basecamp-fe-v1.1.0) (2026-03-03)
+
+
+### Features
+
+* **basecamp:** security hardening for 2FA and basecamp-fe auth ([#471](https://github.com/yeti-robotics/polar-edge/issues/471)) ([fa717f9](https://github.com/yeti-robotics/polar-edge/commit/fa717f9a06ebfa0d34d1b065954723cccde84e93))
+
 ## [1.0.1](https://github.com/yeti-robotics/polar-edge/compare/basecamp-fe-v1.0.0...basecamp-fe-v1.0.1) (2026-02-07)